remove useless viewModel and fields in GoalsFragment, organize package goals
