+++ app/src/main/java/skills/future/planer/db/task/TaskData.java
@@ -16,7 +16,7 @@ import skills.future.planer.db.task.enums.category.TaskCategory;
 import skills.future.planer.db.task.enums.priority.Priorities;
 import skills.future.planer.db.task.enums.priority.TimePriority;
 import skills.future.planer.tools.DatesParser;
-import skills.future.planer.ui.goals.MixedRecyclerElement;
+import skills.future.planer.ui.goals.pager.recycler.MixedRecyclerElement;
 
 /**
  * Class encapsulate data for task
