--- app/src/main/java/skills/future/planer/ui/goals/GoalsFragmentViewModel.java
@@ -1,19 +1,0 @@
-package skills.future.planer.ui.goals;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class GoalsFragmentViewModel extends ViewModel {
-
-    private final MutableLiveData<String> mText;
-
-    public GoalsFragmentViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
