+++ app/src/main/java/skills/future/planer/ui/goals/pager/recycler/MixedRecyclerElement.java
@@ -1,4 +1,4 @@
-package skills.future.planer.ui.goals;
+package skills.future.planer.ui.goals.pager.recycler;
 
 public interface MixedRecyclerElement {
 //    Long idInDatabase;
