+++ app/src/main/java/skills/future/planer/ui/tasklist/viewholders/TaskDataViewHolder.java
@@ -18,8 +18,8 @@ import skills.future.planer.db.task.TaskData;
 import skills.future.planer.db.task.enums.priority.Priorities;
 import skills.future.planer.db.task.enums.priority.TimePriority;
 import skills.future.planer.tools.DatesParser;
-import skills.future.planer.ui.goals.ICustomViewHolder;
-import skills.future.planer.ui.goals.MixedRecyclerElement;
+import skills.future.planer.ui.goals.pager.recycler.ICustomViewHolder;
+import skills.future.planer.ui.goals.pager.recycler.MixedRecyclerElement;
 import skills.future.planer.ui.tasklist.Colors;
 
 @Getter
